Created pausable screen to abstract overlays
