+++ saves/quests.json
@@ -53,14 +53,14 @@ quests: [
 		isSecretQuest: false
 		questDialogue: {
 			class: java.util.ImmutableCollections$MapN
-			TupleKey{str='Cow', num=2}: [
-				Yippeee!
-				You completed your Quest!
-			]
 			TupleKey{str='Cow', num=1}: [
 				Welcome to Animal Kingdom!
 				Here let me help with your quest...
 				Press Spacebar!
+			]
+			TupleKey{str='Cow', num=2}: [
+				Yippeee!
+				You completed your Quest!
 			]
 		}
 		currentTaskIndex: 0
