+++ source/core/src/main/com/csse3200/game/overlays/PauseOverlay.java
@@ -5,6 +5,7 @@ import com.csse3200.game.GdxGame;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.input.InputDecorator;
 import com.csse3200.game.screens.MainGameScreen;
+import com.csse3200.game.screens.PausableScreen;
 import com.csse3200.game.services.ServiceLocator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,16 +18,16 @@ import org.slf4j.LoggerFactory;
 
 public class PauseOverlay extends Overlay {
     private static final Logger logger = LoggerFactory.getLogger(PauseOverlay.class);
-    private MainGameScreen mainGameScreen;
+    private PausableScreen screen;
     private GdxGame game;
 
     /**
      * Constructs an overlay instance and initializes its UI components.
      */
-    public PauseOverlay(MainGameScreen mainGameScreen, GdxGame game) {
+    public PauseOverlay(PausableScreen screen, GdxGame game) {
         super(OverlayType.PAUSE_OVERLAY);
         logger.debug("Initialising PauseOverlay");
-        this.mainGameScreen = mainGameScreen;
+        this.screen = screen;
         this.game = game;
         createUI();
     }
@@ -63,7 +64,7 @@ public class PauseOverlay extends Overlay {
         Stage stage = ServiceLocator.getRenderService().getStage();
         Entity ui = new Entity();
         super.add(ui);
-        ui.addComponent(new PauseDisplay(mainGameScreen, game)).addComponent(new InputDecorator(stage, 10));
+        ui.addComponent(new PauseDisplay(screen, game)).addComponent(new InputDecorator(stage, 10));
         ServiceLocator.getEntityService().register(ui);
     }
 }
