+++ source/core/src/main/com/csse3200/game/overlays/QuestOverlay.java
@@ -4,6 +4,7 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.input.InputDecorator;
 import com.csse3200.game.screens.MainGameScreen;
+import com.csse3200.game.screens.PausableScreen;
 import com.csse3200.game.services.ServiceLocator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,15 +12,15 @@ import org.slf4j.LoggerFactory;
 /** Represents an overlay that displays quest-related information. */
 public class QuestOverlay extends Overlay {
     private static final Logger logger = LoggerFactory.getLogger(QuestOverlay.class);
-    private MainGameScreen mainGameScreen;
+    private PausableScreen screen;
 
 
     /**
      * Constructs a new QuestOverlay.
      */
-    public QuestOverlay(MainGameScreen mainGameScreen) {
+    public QuestOverlay(PausableScreen screen) {
         super(OverlayType.QUEST_OVERLAY);
-        this.mainGameScreen = mainGameScreen;
+        this.screen = screen;
         logger.debug("Initialising QuestOverlay");
         createUI();
     }
@@ -51,7 +52,7 @@ public class QuestOverlay extends Overlay {
         Stage stage = ServiceLocator.getRenderService().getStage();
         Entity ui = new Entity();
         super.add(ui);
-        ui.addComponent(new QuestDisplay(mainGameScreen)).addComponent(new InputDecorator(stage, 10));
+        ui.addComponent(new QuestDisplay(screen)).addComponent(new InputDecorator(stage, 10));
         ServiceLocator.getEntityService().register(ui);
     }
 }
